email_lib: merge MJSON_MIME_* with mime_type enum class

+++ tests/jsontest.cpp
@@ -4,7 +4,7 @@
 
 static void enx(MJSON_MIME *mi, void *q)
 {
-	printf("this=%p type=%d id=%s\n", mi, mi->mime_type, mi->id);
+	printf("this=%p type=%u id=%s\n", mi, static_cast<unsigned int>(mi->mime_type), mi->id);
 }
 
 static char tdata1[] =
